% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health-parse-functions.R
\name{get_data_long}
\alias{get_data_long}
\title{Transform text data (date, key-value pairs) into a long-format data frame.}
\usage{
get_data_long(dat_txt, sep = ";")
}
\arguments{
\item{dat_txt}{Character vector that contains date, time, and one or more
key-value pairs.}

\item{sep}{Character vector of length 1 that specifies which separator
is used to separate the key-value pairs from each other. Default is ";".}
}
\value{
A tibble with three columns: timestamp, key, value. Both key
and value columns are of type character, hence for the value to be used
as numerical data, filtering all non-numeric data and conversion are
still necessary.
}
\description{
Transform text data (date, key-value pairs) into a long-format data frame.
}
\examples{
dat <- c("2001-11-10; 11:00; arzt=OA Dr. Vorname Nachname; Anaesthesist=Dr. Vorname Nachname; Instrumentarin=DGKS Vorname Nachname, DGKP Vorname Nachname; what=some (long) text with semicolons; and other stuff / like slashes, commas, question ? marks, etc.;",
"2001-10-11; 19:40; weight=92.8kg;",
"2001-11-12; 19:30; weight=93.1kg;",
"2001-11-13; 10:00; what=zustand; dauer=5d, leicht kraenklich, husten, schnupfen (!), leichte temperatur (37.3)")
get_data_long(dat)
}
