% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health-parse-functions.R
\name{get_paths}
\alias{get_paths}
\title{Get paths needed for script}
\usage{
get_paths()
}
\value{
List with elements
\code{base} (path of the current project),
\code{r} (path containing r script files),
\code{dat} (path containing the data files),
}
\description{
Get paths needed for script
}
\examples{
  path <- get_paths()
}
