% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health-parse-functions.R
\name{get_subkey_value_mean}
\alias{get_subkey_value_mean}
\title{Extract (mean) values for sub-keys.}
\usage{
get_subkey_value_mean(value, subkey, key_sep = ",", keyvalue_sep = ":",
  vecsep = "/")
}
\arguments{
\item{value}{The value extracted from the original key-value pair.}

\item{subkey}{The subkey for which the data should be extracted.}

\item{key_sep}{The separator that separates the subkey-value-pairs
from each other.}

\item{keyvalue_sep}{The separator that separates the subkey from the value(s)
of that subkey.}

\item{vecsep}{The separator that separates the individual values for each
subkey value.}
}
\value{
A mean value for all values of the subkey, or \code{NA}.
}
\description{
Takes the (vector of) values that contain subkey-value-pairs
of the (default) form:
\code{(subkey1: val1/val2/val3, subkey2: val1/val2/val3)}
and extracts the mean value for a specified subkey.
}
\examples{
#value <- get_value_text(dat_txt, key = "caliper")
#value <- value[357:360]
#subkey <- "brust-li"
dat <- c("2018-03-23; 20:30; caliper = (brust-li: 14/12/11, brust-re: 12/13/13, bauch-li: 25/25/25, bauch-re: 26/26/25, bein-li: 15/15/15, bein-re: 24/23/26);",
  "2018-03-29; 19:00; weight = 90.1kg;", "2018-03-30; 21:00; weight = 89.3kg; note = nach Laufen;",
  "2018-03-30; 21:00; caliper = (brust-li: 12/13/12, brust-re: 12/13/13, bauch-li: 28/29/29, bauch-re: 24/21/28, bein-li: 14/16/14, bein-re: 22/22/21);")
get_subkey_value_mean(
  get_value_text(dat, key = "caliper"),
  subkey = "brust-re")
}
