% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health-parse-functions.R
\name{get_value_num}
\alias{get_value_num}
\title{Get numeric value for a specific key}
\usage{
get_value_num(dat_txt, key, sep = ";")
}
\description{
The function will only return the first value with
that key for a specific date (i.e., multiple identical keys for the same
timestamp will be ignored.). If key is not found, NA will be returned
for that line of the data.
}
