% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health-parse-functions.R
\name{read_healthfiles}
\alias{read_healthfiles}
\title{Read health files.}
\usage{
read_healthfiles(path, filename_regex = "^health-[0-9]{4}\\\\.txt")
}
\arguments{
\item{path}{Path where the files reside.}

\item{filename_regex}{Regular expression that is used to match the
files that should be parsed}
}
\value{
A character vector containting the text in the files (which
usually contains a date, a time, and one or more key-value pairs).
}
\description{
All files that correspond to the regular expression will be read and
post-processed (removing blank lines, trimmed).
}
\examples{
dat_txt <- read_healthfiles("/Users/name/dat_dir")
}
