% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tskeyvalparser-package.R
\docType{package}
\name{tskeyvalparser}
\alias{tskeyvalparser}
\alias{tskeyvalparser-package}
\title{\code{tskeyvalparser} package}
\description{
Parses key-value pairs in a time-series text file.
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
