% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health-parse-functions.R
\name{calc_bodyfat}
\alias{calc_bodyfat}
\title{Calculate body fat from caliper measurements.}
\usage{
calc_bodyfat(value, age, key_sep = ",", keyvalue_sep = ":",
  vecsep = "/")
}
\arguments{
\item{value}{The value extracted from the original key-value pair.}

\item{age}{Age of the person that the caliper measurements were taken of
(at the time of taking the measurements).}

\item{key_sep}{The separator that separates the subkey-value-pairs
from each other.}

\item{keyvalue_sep}{The separator that separates the subkey from the value(s)
of that subkey.}

\item{vecsep}{The separator that separates the individual values for each
subkey value.}
}
\description{
Using the 3-Falten-Formel by Jackson & Pollock, for men.
}
\references{
\url{https://de.wikipedia.org/wiki/Calipometrie}
\url{Jackson, Pollock: Generalized equations for predicting body density of women. In: British Journal of Nutrition. Nr.40, Oktober 1978, S.497–504 (englisch)}
}
